+++ Raytracer/Light.cpp
@@ -6,7 +6,7 @@ Light::Light()
 	intensity = 1;
 }
 
-Light::Light(glm::vec3 color, double intensity)
+Light::Light(glm::vec3 color, float intensity)
 {
 	this->color = color;
 	this->intensity = intensity;
