+++ Raytracer/Plane.h
@@ -9,7 +9,7 @@ public:
 	glm::vec3 normal;
 
 	Plane();
-	Plane(glm::vec3 color, double albedo, glm::vec3 position, glm::vec3 normal);
+	Plane(glm::vec3 color, float albedo, glm::vec3 position, glm::vec3 normal);
 	~Plane();
 
 	virtual Hit intersect(const Ray &ray);
